+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceConfiguration.kt
@@ -8,5 +8,5 @@ data class YospaceConfiguration(
     val connectTimeout: Int = 25_000,
     val requestTimeout: Int = 25_000,
     val liveInitialisationType: YospaceLiveInitialisationType = YospaceLiveInitialisationType.DIRECT,
-    val isDebug: Boolean = false
+    val debug: Boolean = false
 )